Reformat and optimise loading of constants for A64.
Previously we used `opnd_create_reg64` to load 64-bit constants in A64, but the
name was a little misleading as this function actually generates insturctions
and loads a register.  To be more explicit, I've renamed it to
`instr_create_reg64` and reordered the parameters, in keeping with other
instruction insertion methods.  I also optimised it to avoid any unnecessary
instructions that load zero. It could potentially be optimised further using
MOVN instructions, but I don't consider this likely to provide any benefit in
practice.