Rewrite stack profiling storage in dyclient.
The stack counters were previously stored in a map data structure separately
from the main CFG table.  This is actually unnecessary, we can store it in the
main CFG table and pass a pointer to the stored counter around.  This requires a
tiny amount of extra processing in the exit event to handle the case where
blocks overlap.  On the other hand, this means we no longer need to convert
addresses when a module is unloaded.