Fix multi-target rule in Makefile.
The Makefile rule for the DynamoRIO client was using multiple targets, but this
is interpreted by make as multiple independent rules, rather than the intended
interpretation that a signle rule generates two targets. This led to make being
invoked twice redundantly rather than once, possibly causing subtle errors with
file locking. To portably fix, I've instead corrected it to pass a target so
that each sub-make only makes one binary. A better fix would be a single rule
that generates two targets, but this is not portably supported by make.