Fix bugs in handling of JUMP_MBRs and AArch64 MBRs.
For reasons unknown to me it is sometimes the case that the mbr (indirect branch)
target addresses visible to the client are not within the bounds of a loaded
module, despite the fact that the program works fine. We have a system to
correct for this (missed_mbr_table) that notes the next translated block and
assumes the last mbr targeted this block. This system had a bug when used with
the JUMP_MBR optimisation (where repeated mbrs to the same target are optimised
to remove a clean call), as the wrong target address would be used (the
uncorrected one).  Separately, we had included a fix for x86 to remember the
translations from observed pcs to corrected pcs but had not ported this to
Aarch64. This patch fixes both bugs by moving the missed_mbr_table into the
block information and supporting it on both Aarch64 and x86.