Bugfix accounting function samples within loop.
The detection of function stack samples within a loop was slightly incorrect in
that it accumulated samples into loops containing the return address of the
function call, rather than loops containing the call itself. This would make a
difference in the (rare) case where the first block of a loop happened to be the
return of some function call that was not a member of the loop e.g. in code
like:

```c
    func_call();
    do {
       // some loop
    } while (/* condition */);
```