Fix for some instances of incorrect sample address.
For reasons unknown to me, on x86 perf sometimes reports a sample address that
is not within the bounds of a loaded application module, and yet the stack
information associated with that sample is correct and fine. This patch means
that in such cases the top of stack sample address will be used instead of the
sample address, resulting in better output with samples assigned to actual
application addresses. There is no change if perf's output was normal.