Bugfix loop source lines not being computed.
Loop source code ranges were not being properly computed in most cases due to an
overzealous null pointer check. The code was checking if this block has an
inlined function name, and more-or-less refusing to consdier the line numbers if
this was not present. This bug seems to have been around in every release
verison of optiwise.