Improve loop line heuristics.
This patch improves the heuristics used to guess which source line numbers are
associated with a given loop. Previously the heuristic detected various keywords
such as 'for' in source code lines, but without checking for whole words meaning
it could be tricked by variable names such as 'before'. The heuristic also
assumed that most functions would only have a single source filename associated
with them, but in practice this is found to be wrong, and so now the heuristic
prefers the file with the most unique lines associated with it in the event of
multiple files, rather than the first alphabetically. There are still known
issues with this heuristic, hopefully future patches improve it further still.