Add --skip flag to 'optiwise run'
This flag supports skipping individual subcommands of optiwise run, occasionally
useful if a user wishes to reuse a previous result. For example, 'optiwise
disassemble' can be quite slow, and if the underlying binary is not changing
there is no need to repeatedly disassemble it. Hence the user could pass
'optiwise run --skip disassemble -- <PROGRAM>' to run everything except the
disassemble step, reusing a previous disassemble result.