Add --perf-event flag to optiwise check,run,sample.
The new --perf-event flag allows overriding the builtin argument to 'perf
record' which is necessary on some systems where the default value of
'cpu-cyles:uP' causes trouble. Note that 'optiwise analyze' currently parses
specific events by name, so it's likely that using this option with an event
other than 'cpu-cycles' this will lead to empty analysis results, but this can
be addressed in future patches to optiwise analyze.