Improve module path handling in analyzer.
Occasionally it has been useful to reuse results of subcommands (disassemble,
sample, count) for 'optiwise analyze' for example on a different machine.
However, the analyzer currently matched executable paths exactly, so would not
produce correct results unless the full executable paths were consistent for the
three subcommands. This patch causes the analyzer to fall back to matching only
the file name part (not the full path) in the event it cannot find a full path
match. This patch also adds warning messages if this occurs, as well as a big
warning if there are unmatched modules with significant number of smaples or
execution counts.