Improve name handling in 'optiwise gui'.
OptiWISE gui used to use the full demangled function name for filenames and
display names of functions. This can be quite long and redundant for C++
programs as it includes namespaces and template arguments, so this patch
shortens it to use more concise (but hopefully still meaningful) naming in most
cases to make the GUI more comfortable to use.