Prevent crashes in longjmp handling in dyclient.
When a longjmp (or equivalent) is detected, the dynamorio client attempts to
fixup the stack by popping frames. In the event that it fails, it previously
just popped one frame statically, but this could rapidly result in underflowing
the stack buffer. This patch instead causes it to push a new frame, which is
then popped, preventing either overflow or underflow (but ultimately not
resolving the issue of the desynchronised stack). This means for example that an
application that uses coroutines will probably not crash, although stac
profiling will certainly fail.

The patch also fixes an unrelated typo in the warning associated with longjmp
detection, and fixes a broken if condition in the exit event handler.