Fix indirect branches through x28 on aarch64. (#17)
A bug existed causing wrong targets to be reported for branches through x28 on
aarch64 in 'optiwise count'. DynamoRIO has a concept of a 'stolen register': a
register that is not available for use by instrumentation code because DynamoRIO
stores a pointer in it. By default this is x28.  Consequently, if we instrument
an instruction that accesses x28, special care must be taken if we wish to
observe the application state of this instruction. In the dyclient, our indirect
branch target instrumentation code read the value of the target register
naively, so if the target was the stolen register, we would observe the wrong
target address.  This patch detects and fixes this case.  I also slightly
refactored temporary register allocation for both architectures to help
alleviate future problems with stolen registers.